github: run workflows for dependabot
currently, we receive two successive actions when dependabot opens a PR, first an `opened` event, then a `labeled` event. the second event cancels the first pipeline, however, we don't re-run tests when a PR is labeled except if the label is `publish-dev-test`. this fixes the issue by removing the cancel for all labels except `publish-dev-test`.