github: run workflows for dependabot (#33)
Currently, we receive two successive events when dependabot opens a PR,
first an `opened` event, then a `labeled` events, since dependabot
always adds the `dependencies` label.

However, the second event cancels the first workflow, since it is part
of the same concurrency group. Since the second event is a `labeled`
event that isn't about the `publish-dev-test` label, tests are not
re-ran in the second workflow. The PR stays stuck waiting for status
checks for build and test, and those check will never complete.

We fix this by extracting the "run on publish-dev-test" logic to a
separate workflow that calls the pre-release workflow. By creating a
separate workflow, this will not impact status checks, since labeled
actions will live in a separate namespace under "Labeled / ".