docker-compose: fix setup to support secrets in newer versions
newer versions of docker require being root in the container to read
secrets. since we're initializing the profiler and reading the APP and
API keys as non root, we no longer have access to these secrets leading
to failure in uploading symbols.

this fixes the issue by using a shell script to launch the profiler,
which will read the secrets while being root.

also move all Dockerfiles to the `docker/` folder with clearer
separation between developement and release dockerfiles.