cli: fix input of DD_HOST_PROFILING_ADDITIONAL_SYMBOL_ENDPOINTS (#143)
* cli: fix input of DD_HOST_PROFILING_ADDITIONAL_SYMBOL_ENDPOINTS

https://github.com/DataDog/dd-otel-host-profiler/pull/125 by updating
go.mod, updated the urfave/cli dependency from 3.0 alpha to 3.3.

while doing so, it changed in a subtle way how functions provided in the
`Action` field are executed:

per https://cli.urfave.org/v3/examples/flags/value-sources/, now the
library considers environment variables and text files as a way to
override the default value set in the code, as opposed to a dedicated
way to set the flag.

however, when a value is passed with those methods, the library doesn't
consider the flag to be "set" which in turn means it's never added to
https://github.com/urfave/cli/blob/e905d7a/command.go#L545 leading to
the Action function never being executed.

we switch the flag to a GenericFlag with its own `Set` implementation
that is equivalent to the old `Action` implementation. this fixes the
issue since `Set` is always executed to assign the value of the flag,
regardless of the source of the flag (default value in code, environment
value or text file).

* cli: add comments to highlight why we implement these receivers

per review comment

* cli: reword error message for consistency

use API and application key as wording, aligned with the DataDog
documentation