fix: disable the use of HTTP/2 when uploading symbols
currently, we use HTTP/2 when available to upload symbols. however, the
symbol intake backend has only started supporting HTTP/2 more recently,
and we've observed significantly slower uploads when using HTTP/2 due to
misconfigured window sizes on the server-side.

since most of the symbol upload traffic uses datadog-ci, which doesn't
support HTTP/2 (due to the usage of axios), we decided to disable HTTP/2
on the client to align with datadog-ci, and the majority of the symbol
intake traffic.