fix(symbol_uploader): reduce parallel uploads (#189)
reduce parallel uploads from 10 to 5. we've observed that a high number
of parallel uploads can cause the profiler to get OOMKilled on startup,
especially when uploading symbols is slow.

this should reduce the risk of getting OOMKills, while keeping
sufficient paralellism for the uploads to be fast.

the number chosen (5) is slightly random: in a scenario where symbols
can be in order of magnitude of ~100MiB in size, this allows us to cap
memory usage to ~500MiB on startup.