fix(symbol_upload): disable the use of HTTP/2 when uploading symbols (#190)
* fix: disable the use of HTTP/2 when uploading symbols

currently, we use HTTP/2 when available to upload symbols. however, the
symbol intake backend has only started supporting HTTP/2 more recently,
and we've observed significantly slower uploads when using HTTP/2 due to
misconfigured window sizes on the server-side.

since most of the symbol upload traffic uses datadog-ci, which doesn't
support HTTP/2 (due to the usage of axios), we decided to disable HTTP/2
on the client to align with datadog-ci, and the majority of the symbol
intake traffic.

* fix: tests when local objcopy doesn't have zstd support

on workspaces, objcopy has an older version with no zstd support, this
is making all tests fail by default.

this changes the expectation in tests to depend on the output of the
check for objcopy zstd support.