chore: remove main branch check (#203)
since v4.0.0, actions/checkout only checks out the specific ref fetched,
in our case the tag from which we're building a release.

as a consequence, it doesn't fetch the main branch, and does not know on
which ref origin/main is. this is making the check tag step fail
consistently.

since the release workflow is only triggered on tag push, remove the
check to fix the issue.