feat(reporter): default service to Comm (#205)
while looking at profiles, there seems to be a lot of very-short lived
processes from cilium, ctr, or crictl, for which traces contain a valid
Comm (since it is set as the profile root frame), but service is not
detected.

default the service name to `Comm` to further increase the number of
process for which the service is successfully detected. in theory, comm
should always be equivalent to path.Base(execPath).