feat: Add frame type to pprof profile
This commit adds the frame type to the pprof profile, to do so it uses
the system_name field of pprof.Function.
It also always emits a Line for non-native frames. When the function name
is empty (this can happen for abort frames and kernel/interpreted frames
symbolication failures), we use "Unknown Function" as function name.